feat(table-card): add TableCard usage for Table stories #735
